requirements.txt
albumentations==1.3.1
click==8.1.6
clip==1.0
gradio==4.19.2
matplotlib==3.8.2
MedCLIP==0.0.3
monai==1.2.0
mpire==2.8.1
numpy==1.24.4
openai==1.12.0
opencv_python==4.8.0.74
opencv_python_headless==4.8.0.74
pandas==2.0.3
Pillow==10.2.0
scikit_learn==1.3.0
seaborn==0.13.2
tencentcloud_sdk_python==3.0.992
torch==2.0.1+cu118
torchmetrics==1.2.0
torchvision==0.15.2+cu118
tqdm==4.65.0
transformers==4.33.1
wandb==0.16.3
